backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
llama-index==0.12.6
llama-index-llms-groq==0.3.1
llama-index-embeddings-huggingface==0.4.0
llama-index-readers-file==0.4.1
llama-index-core==0.12.6
python-dotenv==1.0.1
pydantic>=2.8.0
groq==0.4.2